서버투서버 통신오류 때문에 semesterId 하드코딩으로 바꿔놓음

+++ src/main/java/com/green/muziuniv_be_user/application/department/model/DepartmentGetRes.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DepartmentGetRes {
-    private int deptId;
+    private Long deptId;
     private String deptName;
     private String status;
     private int headProfId;
